Moved utils to one file (useThemeDetector and getDarkTheme) #5
